Add a loading wheel and freeze frame before awaiting logs in AICamera
